Broken down the continuous graph in subfiles
