+++ scrutiny/gui/dashboard_components/continuous_graph/continuous_graph_component.py
@@ -10,16 +10,12 @@ from datetime import datetime
 import functools
 import math
 from pathlib import Path
-import logging
 from dataclasses import dataclass
-import re
-import os
 
-from PySide6.QtGui import QPainter, QFontMetrics, QFont, QColor, QContextMenuEvent, QKeyEvent
-from PySide6.QtWidgets import (QHBoxLayout, QSplitter, QWidget, QVBoxLayout,  QMenu, QSizePolicy,
-                               QPushButton, QFormLayout, QSpinBox, QGraphicsItem, QStyleOptionGraphicsItem,
-                               QLineEdit, QCheckBox, QGroupBox, QMessageBox)
-from PySide6.QtCore import Qt, QItemSelectionModel, QPointF, QTimer, QRectF, QRect
+from PySide6.QtGui import QContextMenuEvent, QKeyEvent
+from PySide6.QtWidgets import (QHBoxLayout, QSplitter, QWidget, QVBoxLayout,  QMenu,
+                               QPushButton, QFormLayout, QSpinBox, QLineEdit)
+from PySide6.QtCore import Qt, QItemSelectionModel, QPointF, QTimer, QRectF
 
 from scrutiny import sdk
 from scrutiny import tools
@@ -27,7 +23,6 @@ from scrutiny.gui import assets
 from scrutiny.gui.app_settings import app_settings
 from scrutiny.gui.tools import prompt
 from scrutiny.gui.tools.invoker import InvokeQueued, InvokeInQtThread
-from scrutiny.gui.tools.min_max import MinMax
 from scrutiny.gui.core.definitions import WidgetState
 from scrutiny.gui.core.watchable_registry import WatchableRegistryNodeNotFoundError, ValueUpdate
 from scrutiny.gui.widgets.feedback_label import FeedbackLabel
@@ -37,402 +32,13 @@ from scrutiny.gui.dashboard_components.common.export_chart_csv import export_cha
 from scrutiny.gui.dashboard_components.common.base_chart import (
     ScrutinyLineSeries, ScrutinyValueAxisWithMinMax, ScrutinyChartCallout, ScrutinyChartView, ScrutinyChart,
     ScrutinyChartToolBar)
-from scrutiny.gui.dashboard_components.continuous_graph.decimator import GraphMonotonicNonUniformMinMaxDecimator
+from scrutiny.gui.dashboard_components.continuous_graph.csv_logging_menu import CsvLoggingMenuWidget
+from scrutiny.gui.dashboard_components.continuous_graph.realtime_line_series import RealTimeScrutinyLineSeries
+from scrutiny.gui.dashboard_components.continuous_graph.graph_statistics import GraphStatistics
 from scrutiny.gui.core.preferences import gui_preferences
-from scrutiny.sdk.listeners.csv_logger import CSVLogger, CSVConfig
-from scrutiny.tools.profiling import VariableRateExponentialAverager
+from scrutiny.sdk.listeners.csv_logger import CSVLogger
 
-from typing import Dict, Any, Union, List, Optional, cast, Set, Generator
-
-class CsvLoggingMenuWidget(QWidget):
-    _chk_enable:QCheckBox
-    _txt_folder:QLineEdit
-    _txt_filename_pattern:QLineEdit
-    _spin_max_line_per_file:QSpinBox
-    _gb_content:QGroupBox
-
-    def __init__(self, parent:QWidget):
-        super().__init__(parent)
-        layout = QVBoxLayout(self)
-        layout.setContentsMargins(0,0,0,0)
-
-        self._chk_enable = QCheckBox("Log to CSV", self)
-        self._gb_content = QGroupBox(self)
-        self._gb_content.setVisible(False)
-        layout.addWidget(self._chk_enable)
-        layout.addWidget(self._gb_content)
-        self._txt_folder = QLineEdit(self)
-        self._txt_filename_pattern = QLineEdit(self)
-        self._spin_max_line_per_file = QSpinBox(self)
-        self._spin_max_line_per_file.setMinimum(1000)
-        self._spin_max_line_per_file.setMaximum(1000000)
-        self._spin_max_line_per_file.setValue(10000)
-        self._btn_browse = QPushButton("...", self)
-        self._txt_folder.setSizePolicy(QSizePolicy.Policy.Expanding, QSizePolicy.Policy.Preferred)
-        self._btn_browse.setSizePolicy(QSizePolicy.Policy.Minimum, QSizePolicy.Policy.Preferred)
-        self._btn_browse.setMinimumWidth(20)
-        self._btn_browse.setMaximumWidth(40)
-
-        self._btn_browse.clicked.connect(self._browse_clicked_slot)
-
-        filebrowse_line = QWidget()
-        filebrowse_layout = QHBoxLayout(filebrowse_line)
-        filebrowse_layout.setContentsMargins(0,0,0,0)
-        filebrowse_layout.addWidget(self._txt_folder)
-        filebrowse_layout.addWidget(self._btn_browse)
-
-        gb_layout = QFormLayout(self._gb_content)
-
-        gb_layout.addRow("Folder", filebrowse_line)
-        gb_layout.addRow("File prefix", self._txt_filename_pattern)
-        gb_layout.addRow("Lines/file", self._spin_max_line_per_file)
-
-        self._chk_enable.checkStateChanged.connect(self._check_state_changed_slot)
-
-
-    def _check_state_changed_slot(self, state:Qt.CheckState) -> None:
-        if state == Qt.CheckState.Checked:
-            self._gb_content.setVisible(True)
-        else:
-            self._gb_content.setVisible(False)
-
-    def _browse_clicked_slot(self) -> None:
-        actual_folder:Optional[Path] = None # USe last save dir if None
-        if os.path.isdir(self._txt_folder.text()):
-            actual_folder = Path(os.path.normpath(self._txt_folder.text()))
-            actual_folder = actual_folder.absolute()
-
-        folder = prompt.get_save_folderpath_from_last_save_dir(self, "Select a folder", save_dir=actual_folder)
-        if folder is not None:
-            self._txt_folder.setText(str(folder))
-    
-    def require_csv_logging(self) -> bool:
-        return self._chk_enable.isChecked()
-
-    def validate(self) -> None:
-        folder = self._txt_folder.text()
-        if len(folder) == 0:
-            raise ValueError("No folder selected")
-        filename_pattern = self._txt_filename_pattern.text()
-        if len(filename_pattern) == 0:
-            raise ValueError("No filename prefix provided")
-
-        valid_filename = re.compile(r"^[A-Za-z0-9\._\-\(\)]+$")
-        if not valid_filename.match(filename_pattern):
-            raise ValueError("Invalid characters in filename")
-        
-        folder = os.path.normpath(folder)
-        if not os.path.isabs(folder):
-            folder = os.path.normpath(os.path.abspath(folder))
-
-        if not os.path.isdir(folder):
-            raise FileNotFoundError(f"Folder {folder} does not exist")
-
-        self._txt_folder.setText(folder)
-    
-    def check_conflicts(self) -> bool:
-        self.validate()
-        
-        folder = Path(self._txt_folder.text())
-        filename = self._txt_filename_pattern.text()
-
-        conflicting_files = list(CSVLogger.get_conflicting_files(folder, filename))
-        if len(conflicting_files) == 0:
-            return True
-        
-        MAX_NAME_DISPLAY = 3
-        msgbox_text = f"There are {len(conflicting_files)} existing files that may conflict with the given folder/name combination.\n"
-        msgbox_text += '\n'.join([f"  - {file.name}" for file in conflicting_files[:MAX_NAME_DISPLAY] ]) + '\n'
-        if len(conflicting_files) > MAX_NAME_DISPLAY:
-            diff = len(conflicting_files) - MAX_NAME_DISPLAY
-            msgbox_text += f'  - And {diff} others\n'
-
-        msgbox_text += '\n Do you want to delete them?'
-        msgbox = QMessageBox(self)
-        msgbox.setIcon(QMessageBox.Icon.Warning)
-        msgbox.setWindowTitle("Filename conflict")
-        msgbox.setText(msgbox_text)
-        msgbox.setStandardButtons(QMessageBox.StandardButton.No | QMessageBox.StandardButton.Yes)
-        msgbox.setDefaultButton(QMessageBox.StandardButton.No)
-
-        msgbox.setModal(True)
-        reply = msgbox.exec()
-
-        if reply == QMessageBox.StandardButton.Yes:
-            for file in conflicting_files:
-                os.unlink(file)
-            return True
-        elif reply == QMessageBox.StandardButton.No:
-            return False
-        else:
-            raise NotImplementedError("Unsupported response")
-    
-    def make_csv_logger(self, logging_logger:Optional[logging.Logger] = None)  -> CSVLogger:
-        # Validation happens inside the constructor
-        self.validate()
-
-        csv_config = CSVConfig(
-            delimiter=',',
-            newline='\n'
-        )
-
-        return CSVLogger(
-            folder = self._txt_folder.text(),
-            filename = self._txt_filename_pattern.text(),
-            lines_per_file = self._spin_max_line_per_file.value(),
-            datetime_format=gui_preferences.global_namespace().long_datetime_format(),
-            csv_config=csv_config,
-            convert_bool_to_int=True,
-            logger=logging_logger,
-            file_part_0pad=4
-        )
-
-class RealTimeScrutinyLineSeries(ScrutinyLineSeries):
-    """Extension of a LineSeries that is meant to display data in real time.
-    It has support for decimation and some fancy tricks to kep tracks of min/max
-    value with minimal CPU computation"""
-
-    _decimator:GraphMonotonicNonUniformMinMaxDecimator
-    """The decimator that keeps the whole dataset and also provides a decimated version of it"""
-    _x_minmax:MinMax
-    """Min/Max trackerfor the X values"""
-    _y_minmax:MinMax
-    """Min/Max trackerfor the Y values"""
-    _dirty:bool
-    """Flag indicating that the output of the decimator has new data ready to be flushed to the chart"""
-
-    def __init__(self, *args:Any, **kwargs:Any) -> None:
-        super().__init__(*args, **kwargs)
-        self._decimator = GraphMonotonicNonUniformMinMaxDecimator()
-        self._x_minmax = MinMax()
-        self._y_minmax = MinMax()
-        self._dirty = False
-
-    def set_x_resolution(self, resolution:float) -> bool:
-        """Set the X width used by the decimator. All the points within a moving window of this size will be clustered together"""
-        changed = self._decimator.set_x_resolution(resolution)
-        if changed:            
-            self._dirty = True
-        return changed
-    
-    def decimation_factor(self) -> float:
-        """Provides an estimation of the decimation factor"""
-        return self._decimator.decimation_factor()
-    
-    def count_decimated_points(self) -> int:
-        """Count the number of points at the output of the decimator"""
-        return len(self._decimator.get_decimated_buffer())
-
-    def count_all_points(self) -> int:
-        """Count the number of points at the input of the decimator (full dataset)"""
-        return len(self._decimator.get_input_buffer())
-
-    def add_point(self, point:QPointF) -> int:
-        """Adds a point to the decimator input. New output points may or may not be available after
-
-        :param point: The point to add
-        :return: The number of new points available at the output. May be bigger than 1
-        """
-        n = self._decimator.add_point(point)    # Can have 2 points out for a single in (min/max)
-        if n > 0:
-            # New data available at the output. update the min/max in real time for axis autorange
-            start_index = len(self._decimator.get_decimated_buffer())-n     # Avoid negative slice. Not all container supports it
-            for p in self._decimator.get_decimated_buffer()[start_index:]:
-                self._y_minmax.update(p.y())
-                self._x_minmax.update_max(p.x())
-            self._dirty=True
-        return n
-    
-    def delete_up_to_x_without_flushing(self, x:float) -> None:
-        """Delete both input and output data up to a value of X specified. Assumes a monotonic X axis
-        
-        :param x: The minimum X value allowed. Any points with a X value smaller than this will get deleted
-        
-        """
-        in_deleted, out_deleted = self._decimator.delete_data_up_to_x(x)
-        if out_deleted > 0: # That affects the visible graph
-            self._dirty = True
-    
-    def get_last_x(self) -> Optional[float]:
-        """Return the most recent point from the input buffer"""
-        buffer = self._decimator.get_input_buffer()
-        if len(buffer) == 0:
-            return None
-        return buffer[-1].x()
-
-    def get_first_x(self) -> Optional[float]:
-        """Return the oldest point from the input buffer"""
-        buffer = self._decimator.get_input_buffer()
-        if len(buffer) == 0:
-            return None
-        return buffer[0].x()
-    
-    def get_last_decimated_x(self) -> Optional[float]:
-        """Return the most recent point fromthe output buffer (decimated buffer)"""
-        buffer = self._decimator.get_decimated_buffer()
-        if len(buffer) == 0:
-            return None
-        return buffer[-1].x()
-
-    def get_first_decimated_x(self) -> Optional[float]:
-        """Return the oldest point fromthe output buffer (decimated buffer)"""
-        buffer = self._decimator.get_decimated_buffer()
-        if len(buffer) == 0:
-            return None
-        return buffer[0].x()
-
-    def flush_decimated(self) -> None:
-        """Copy the decimator output buffer (decimated) into the chart buffer for display"""
-        self.replace(self._decimator.get_decimated_buffer())
-        self._dirty = False
-    
-    def flush_full_dataset(self) -> None:
-        """Copy the decimator input buffer (full dataset) into the chart buffer for display"""
-        self.replace(self._decimator.get_input_buffer())
-
-    def is_dirty(self) -> bool:
-        return self._dirty
-
-    def stop_decimator(self) -> None:
-        """Stops the decimator, Any input pointed being held for decimation will be moved to the output"""
-        self._decimator.force_flush_pending()
-
-    def recompute_minmax(self) -> None:
-        """Recompute the min/max values of the whole dataset"""
-
-        # Since the decimator always keeps the min/max of a cluster, we can safely use the decimated buffer to 
-        # accurately compute the min/max of the whole dataset. We need to take in account the few points at the
-        # end of the input buffer not yet moved to the output
-        decimated_buffer = self._decimator.get_decimated_buffer()       # The output buffer
-        unprocessed_inputs = self._decimator.get_unprocessed_input()    # The few most recent points at the input buffer
-        
-        self._x_minmax.clear()
-        self._y_minmax.clear()
-
-        # Tests have shown that it is ~2x faster to iterate twice (1 to extract the right value, second to run the min/max function) than
-        # iterate once with a key specifier
-        self._x_minmax.update_from_many([p.x() for p in decimated_buffer])  
-        self._y_minmax.update_from_many([p.y() for p in decimated_buffer])
-
-        self._x_minmax.update_from_many([p.x() for p in unprocessed_inputs])
-        self._y_minmax.update_from_many([p.y() for p in unprocessed_inputs])
-
-    def x_min(self) -> Optional[float]:
-        """The smallest X value in the whole dataset"""
-        return self._x_minmax.min()
-    
-    def x_max(self) -> Optional[float]:
-        """The largest X value in the whole dataset"""
-        return self._x_minmax.max()
-    
-    def y_min(self) -> Optional[float]:
-        """The smallest Y value in the whole dataset"""
-        return self._y_minmax.min()
-    
-    def y_max(self) -> Optional[float]:
-        """The largest Y value in the whole dataset"""
-        return self._y_minmax.max()
-
-class GraphStatistics:
-    class Overlay(QGraphicsItem):
-        _stats:"GraphStatistics"
-        _bounding_box:QRectF
-        _text_rect:QRectF
-        _font:QFont
-        _text:str
-
-        def __init__(self, parent:Optional[QGraphicsItem], stats:"GraphStatistics") -> None:
-            super().__init__(parent)
-            self._stats = stats
-            self._bounding_box = QRectF()
-            self._text_rect = QRectF()
-            self._font  = QFont()
-            self._text = ""
-            self.setZValue(11)
-
-        def update_content(self) -> None:
-            self._make_text()
-            self._compute_geometry()
-
-        def _make_text(self) -> None:
-            refresh_rate_str = "N/A"
-            if self._stats.repaint_rate.is_enabled():
-                refresh_rate_str = "%0.1f/sec" % self._stats.repaint_rate.get_value()
-            opengl_enabled_str = "Enabled" if self._stats.opengl else "Disabled"
-            lines = [
-                "Decimation: %0.1fx" % self._stats.decimation_factor,
-                "Visible points: %d/%d" % (self._stats.visible_points, self._stats.total_points),
-                "OpenGL: %s" % opengl_enabled_str,
-                "Refresh rate: %s" % refresh_rate_str 
-            ]
-            self._text =  '\n'.join(lines)
-
-        def _compute_geometry(self) -> None:
-            self.prepareGeometryChange()
-            metrics = QFontMetrics(self._font)
-            self._text_rect = QRectF(metrics.boundingRect(QRect(0, 0, 200, 100), Qt.AlignmentFlag.AlignLeft, self._text))
-            self._bounding_box = QRectF(self._text_rect.adjusted(0,0,0,0))
-
-        def boundingRect(self) -> QRectF:
-            return self._bounding_box
-
-        def paint(self, painter: QPainter, option: QStyleOptionGraphicsItem, widget: Optional[QWidget]=None) -> None:
-            painter.fillRect(self._text_rect, QColor(0xC8, 0xE8, 0xFF, 50))
-            painter.setPen(QColor(0,0,0))
-            painter.drawText(self._text_rect, self._text)
-
-
-    visible_points:int
-    """Number of points drawn on the chart"""
-    total_points:int
-    """Number of points stored in memory that could go on the graph (does not include points so old that they get removed)"""
-    decimation_factor:float
-    """Ratio of visible points/total points"""
-    opengl:bool
-    """Says if OpenGL is activated"""
-    repaint_rate:VariableRateExponentialAverager
-    """Estimated repaint rate in number of repaint/sec"""
-    _overlay:Overlay
-    """The GraphicItem overlay being displayed on top of the chart"""
-    _allow_show_overlay:bool
-    """A flag enabling/disabling the overlay"""
-
-    def __init__(self, draw_zone:Optional[QGraphicsItem] = None) -> None:
-        self._allow_show_overlay = True
-        self._overlay = self.Overlay(draw_zone, self)
-        self.repaint_rate = VariableRateExponentialAverager(time_estimation_window=0.2, tau=1, near_zero=0.01)
-        self.clear()
-
-    def clear(self) -> None:
-        self.opengl = False
-        self.visible_points = 0
-        self.total_points = 0
-        self.decimation_factor = 1
-
-    def show_overlay(self) -> None:
-        if self._allow_show_overlay:
-            self._overlay.show()
-            self._overlay.update_content()
-
-    def hide_overlay(self) -> None:
-        self._overlay.hide()
-
-    def update_overlay(self) -> None:
-        self._overlay.update_content()
-
-    def overlay(self) -> Overlay:
-        return self._overlay
-
-    def allow_overlay(self) -> None:
-        self._allow_show_overlay = True
-        self.show_overlay()
-
-    def disallow_overlay(self) -> None:
-        self._allow_show_overlay = False
-        self.hide_overlay()
-    
-    def is_overlay_allowed(self) -> bool:
-        return self._allow_show_overlay
+from scrutiny.tools.typing import *
 
 @dataclass
 class ContinuousGraphState:
